requirements.txt
unsloth
torch
transformers
trl
datasets
pandas
pyyaml
accelerate
bitsandbytes
peft
xformers
sentencepiece
